refactor(common): small improvements
