+++ core/design-system/src/commonMain/kotlin/com/neoutils/neoregex/core/designsystem/theme/Dimensions.kt
@@ -52,3 +52,6 @@ data class Dimensions(
         x = 56.dp
     )
 )
+
+expect val MinimalButtonHeight: Dp
+expect val TopBarHeight: Dp
